feat: discovery and hpke mod
